Move MLS mocks to a separate module
